refactor walker
