+++ packages/cspell-trie-lib/src/lib/walker/index.ts
@@ -1,0 +1,3 @@
+export * from './walker';
+export * from './hintedWalker';
+export * from './walkerTypes';
