+++ packages/cspell-trie-lib/src/lib/walker/walker.test.ts
@@ -1,6 +1,8 @@
-import { walker, YieldResult, hintedWalker } from './walker';
-import { orderTrie, createTriFromList } from './trie-util';
-import { parseLinesToDictionary } from './SimpleDictionaryParser';
+import { walker } from './walker';
+import type { YieldResult } from './walkerTypes';
+import { hintedWalker } from './hintedWalker';
+import { orderTrie, createTriFromList } from '../trie-util';
+import { parseLinesToDictionary } from '../SimpleDictionaryParser';
 
 describe('Validate Util Functions', () => {
     test('Tests Walker', () => {
