+++ packages/cspell-trie-lib/src/lib/walker/walker.ts
@@ -1,0 +1,45 @@
+import { TrieNode, ChildMap } from '../TrieNode';
+import { CompoundWordsMethod, WalkerIterator, WORD_SEPARATOR, JOIN_SEPARATOR } from './walkerTypes';
+
+/**
+ * Walks the Trie and yields a value at each node.
+ * next(goDeeper: boolean):
+ */
+export function* walker(
+    root: TrieNode,
+    compoundingMethod: CompoundWordsMethod = CompoundWordsMethod.NONE
+): WalkerIterator {
+    const roots: { [index: number]: ChildMap | [string, TrieNode][] } = {
+        [CompoundWordsMethod.NONE]: [],
+        [CompoundWordsMethod.JOIN_WORDS]: [[JOIN_SEPARATOR, root]],
+        [CompoundWordsMethod.SEPARATE_WORDS]: [[WORD_SEPARATOR, root]],
+    };
+
+    function* children(n: TrieNode): IterableIterator<[string, TrieNode]> {
+        if (n.c) {
+            yield* n.c;
+        }
+        if (n.f) {
+            yield* roots[compoundingMethod];
+        }
+    }
+
+    let depth = 0;
+    const stack: { t: string; c: Iterator<[string, TrieNode]> }[] = [];
+    stack[depth] = { t: '', c: children(root) };
+    let ir: IteratorResult<[string, TrieNode]>;
+    while (depth >= 0) {
+        let baseText = stack[depth].t;
+        while (!(ir = stack[depth].c.next()).done) {
+            const [char, node] = ir.value;
+            const text = baseText + char;
+            const goDeeper = yield { text, node, depth };
+            if (goDeeper || goDeeper === undefined) {
+                depth++;
+                baseText = text;
+                stack[depth] = { t: text, c: children(node) };
+            }
+        }
+        depth -= 1;
+    }
+}
